+++ internal/provider/log_binding_resource.go
@@ -40,7 +40,8 @@ type logBindingResourceModel struct {
 	Name        types.String `tfsdk:"name"`
 	SpecID      types.String `tfsdk:"spec_id"`
 	LogstreamID types.String `tfsdk:"logstream_id"`
-	GrokPattern types.String `tfsdk:"grok_pattern"`
+	Pattern     types.String `tfsdk:"pattern"`
+	PatternType types.String `tfsdk:"pattern_type"`
 }
 
 // Configure adds the provider configured client to the resource.
@@ -82,8 +83,12 @@ func (r *logBindingResource) Schema(_ context.Context, _ resource.SchemaRequest,
 				Description: "The specification id.",
 				Required:    true,
 			},
-			"grok_pattern": schema.StringAttribute{
-				Description: "The base_path for this log binding.",
+			"pattern": schema.StringAttribute{
+				Description: "The grok/json pattern for this log binding.",
+				Required:    true,
+			},
+			"pattern_type": schema.StringAttribute{
+				Description: "The pattern type for this log binding. Accepted values: grok, json",
 				Required:    true,
 			},
 			"logstream_id": schema.StringAttribute{
@@ -114,7 +119,8 @@ func (r *logBindingResource) Create(ctx context.Context, req resource.CreateRequ
 	bindingResponse, _, err := r.client.LogBindingsAPI.CreateLogBinding(ctx, r.client.OrgID).LogBindingPostBody(openapiclient.LogBindingPostBody{
 		Name:        plan.Name.ValueString(),
 		SpecId:      plan.SpecID.ValueString(),
-		GrokPattern: base64.StdEncoding.EncodeToString([]byte(plan.GrokPattern.ValueString())),
+		Pattern:     base64.StdEncoding.EncodeToString([]byte(plan.Pattern.ValueString())),
+		PatternType: plan.PatternType.ValueString(),
 		LogstreamId: &logstreamID,
 	}).Execute()
 
@@ -136,14 +142,14 @@ func (r *logBindingResource) Create(ctx context.Context, req resource.CreateRequ
 	plan.Name = types.StringValue(bindingResponse.Name)
 	plan.SpecID = types.StringValue(bindingResponse.SpecId)
 
-	grokPattern, err := base64.StdEncoding.DecodeString(bindingResponse.GrokPattern)
+	pattern, err := base64.StdEncoding.DecodeString(bindingResponse.Pattern)
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Unable to base64 decode the grok pattern",
+			"Unable to base64 decode the pattern",
 			err.Error(),
 		)
 	}
-	plan.GrokPattern = types.StringValue(string(grokPattern))
+	plan.Pattern = types.StringValue(string(pattern))
 
 	if logstreamID != "" {
 		plan.LogstreamID = types.StringValue(bindingResponse.LogstreamId)
@@ -193,23 +199,24 @@ func (r *logBindingResource) Read(ctx context.Context, req resource.ReadRequest,
 
 	// Map response body to model
 	state = logBindingResourceModel{
-		ID:     types.StringValue(bindingResponse.Id),
-		Name:   types.StringValue(bindingResponse.Name),
-		SpecID: types.StringValue(bindingResponse.SpecId),
+		ID:          types.StringValue(bindingResponse.Id),
+		Name:        types.StringValue(bindingResponse.Name),
+		SpecID:      types.StringValue(bindingResponse.SpecId),
+		PatternType: types.StringValue(bindingResponse.PatternType),
 	}
 
 	if logstreamID != "" {
 		state.LogstreamID = types.StringValue(bindingResponse.LogstreamId)
 	}
 
-	grokPattern, err := base64.StdEncoding.DecodeString(bindingResponse.GrokPattern)
+	pattern, err := base64.StdEncoding.DecodeString(bindingResponse.Pattern)
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Unable to base64 decode the grok pattern",
+			"Unable to base64 decode the pattern",
 			err.Error(),
 		)
 	}
-	state.GrokPattern = types.StringValue(string(grokPattern))
+	state.Pattern = types.StringValue(string(pattern))
 
 	// Set refreshed state
 	diags = resp.State.Set(ctx, &state)
@@ -240,8 +247,9 @@ func (r *logBindingResource) Update(ctx context.Context, req resource.UpdateRequ
 		LogBindingPutBody(openapiclient.LogBindingPutBody{
 			Name:        plan.Name.ValueString(),
 			SpecId:      plan.SpecID.ValueString(),
-			GrokPattern: base64.StdEncoding.EncodeToString([]byte(plan.GrokPattern.ValueString())),
+			Pattern:     base64.StdEncoding.EncodeToString([]byte(plan.Pattern.ValueString())),
 			LogstreamId: logstreamId,
+			PatternType: plan.PatternType.ValueString(),
 		})
 
 	// update the log binding
@@ -264,18 +272,19 @@ func (r *logBindingResource) Update(ctx context.Context, req resource.UpdateRequ
 	state := logBindingResourceModel{
 		ID:          types.StringValue(logBindingResponse.Id),
 		Name:        types.StringValue(logBindingResponse.Name),
-		GrokPattern: types.StringValue(logBindingResponse.GrokPattern),
+		Pattern:     types.StringValue(logBindingResponse.Pattern),
 		SpecID:      types.StringValue(logBindingResponse.SpecId),
+		PatternType: types.StringValue(logBindingResponse.PatternType),
 	}
 
-	grokPattern, err := base64.StdEncoding.DecodeString(logBindingResponse.GrokPattern)
+	pattern, err := base64.StdEncoding.DecodeString(logBindingResponse.Pattern)
 	if err != nil {
 		resp.Diagnostics.AddError(
-			"Unable to base64 decode the grok pattern",
+			"Unable to base64 decode the pattern",
 			err.Error(),
 		)
 	}
-	state.GrokPattern = types.StringValue(string(grokPattern))
+	state.Pattern = types.StringValue(string(pattern))
 
 	// update state if logstream id was set
 	if !plan.LogstreamID.IsNull() {
@@ -317,4 +326,21 @@ func (r *logBindingResource) Delete(ctx context.Context, req resource.DeleteRequ
 	}
 
 	tflog.Debug(ctx, "Deleted the log binding resource", map[string]any{"success": true})
+}
+
+func (r logBindingResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
+	var data logBindingResourceModel
+
+	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
+
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if data.PatternType.ValueString() != "json" && data.PatternType.ValueString() != "grok" {
+		resp.Diagnostics.AddError(
+			"Invalid pattern_type",
+			"Accepted values: grok, json",
+		)
+	}
 }